+++ crates/ladfile_builder/readme.md
@@ -1,0 +1,3 @@
+# Language Agnostic Declaration file builder
+
+Contains utilities for building LAD's using `bevy_mod_scripting_core` and `bevy`.
